Add authorization details for data frame analytics jobs
